First pass at moving to ES Module
