+++ lib/config.js
@@ -1,8 +1,12 @@
 'use strict';
 
-const fs = require('fs');
-const Ajv = require('ajv');
-const schema = require('./schema.json');
+import fs from 'fs';
+import Ajv from 'ajv';
+import { dirname } from 'path';
+import path from 'path';
+import { fileURLToPath } from 'url';
+
+const schema = fs.readFileSync(path.resolve(dirname(fileURLToPath(import.meta.url)), './schema.json'));
 
 /**
  * @class
@@ -105,4 +109,4 @@ class Config {
     }
 }
 
-module.exports = Config;
+export default Config;
