+++ lib/error.js
@@ -14,4 +14,4 @@ class TBError extends Error {
     }
 }
 
-module.exports = TBError;
+export default TBError;
