+++ lib/interfaces.js
@@ -1,7 +1,11 @@
 'use strict';
-module.exports = {
-    'file:': require('./interfaces/file'),
-    's3:': require('./interfaces/s3'),
-    'http:': require('./interfaces/http'),
-    'https:': require('./interfaces/http')
+import file from './interfaces/file.js';
+import s3 from './interfaces/s3.js';
+import http from './interfaces/http.js';
+
+export default {
+    'file:': file,
+    's3:': s3,
+    'http:': http,
+    'https:': http
 };
