+++ lib/interfaces/file.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const fs = require('fs/promises');
+import fs from 'fs/promises';
 
 /**
  * @class
@@ -44,4 +44,4 @@ class TileBaseFile {
     }
 }
 
-module.exports = TileBaseFile;
+export default TileBaseFile;
