+++ lib/interfaces/http.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const fetch = require('node-fetch');
+import fetch from 'node-fetch';
 
 /**
  * @class
@@ -44,4 +44,4 @@ class TileBaseHTTP {
     }
 }
 
-module.exports = TileBaseHTTP;
+export default TileBaseHTTP;
