+++ lib/interfaces/s3.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const AWS = require('aws-sdk');
+import AWS from 'aws-sdk';
 const s3 = new AWS.S3();
 
 /**
@@ -45,4 +45,4 @@ class TileBaseS3 {
     }
 }
 
-module.exports = TileBaseS3;
+export default TileBaseS3;
