+++ test/to_tb_complicated.js
@@ -1,11 +1,11 @@
 'use strict';
 
-const test = require('tape');
-const path = require('path');
-const TileBase = require('../tilebase.js');
-const { VectorTile } = require('@mapbox/vector-tile');
-const MBTiles = require('@mapbox/mbtiles');
-const Protobuf = require('pbf');
+import test from 'tape';
+import path from 'path';
+import TileBase from '../tilebase.js';
+import { VectorTile } from '@mapbox/vector-tile';
+import MBTiles from '@mapbox/mbtiles';
+import Protobuf from 'pbf';
 
 test('TileBase#To_TB (complicated)', async (t) => {
     try {
